Create BigInt
This algorithm has the ability to convert any integer to an array of size 20 and it will also do the same conversion with a string. There are mathematical methods as well such as compare that will tell you when two arrays are less than, greater than, or equal to; there is add that will add two integer arrays together; there is mull that will multiply two arrays together.